requirements.txt
boto3==1.29.0
botocore==1.32.0
fastapi==0.103.1
install==1.3.5
ipykernel==6.25.2
lxml==4.9.3
matplotlib-inline==0.1.6
numpy==1.25.2
packaging==23.1
panda==0.3.1
pandas==2.0.3
plotly==5.16.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.22
tabula-py==2.9.0
